Move tests from DiagnosticTests to other test files
Now that `AssertParse` is able to assert for diagnostics, it doesn’t make sense to keep the diagnostic test cases in a separate test file.
